Conflicts:
	CMakeLists.txt
	plugins/welcomepage/kdevwelcomepage.json

+++ kdevplatform/language/duchain/indexedtopducontext.h
@@ -133,7 +133,7 @@ struct IndexedTopDUContextIndexConversion
     }
 };
 
-inline uint qHash(const IndexedTopDUContext& ctx)
+inline size_t qHash(const IndexedTopDUContext& ctx)
 {
     return ctx.index();
 }
